ml_system_design_generator/pyproject.toml
[project]
name = "ml-system-design-generator"
version = "0.1.0"
description = "CLI tool that reads markdown source docs and generates publication-quality ML system design documents in LaTeX"
requires-python = ">=3.10"
dependencies = [
    "ag2[openai]",
    "pydantic>=2.7.0",
    "pyyaml>=6.0",
    "openai>=1.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "chromadb>=0.4.0",
    "requests>=2.31.0",
]

[dependency-groups]
dev = ["pytest>=8.0", "ruff>=0.8"]

[project.scripts]
mlsd = "ml_system_design_generator.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
